fix: Delete Product Images when the Product is deleted from the database
Previously, Product Images were deleted on the deletion of Product from SalesConfig and ProductConfig Screens only. This fix ensures Product Images are deleted on the deletion of Product from the List screens by executing the image deletion process on success of product deletion in the Repository implementation, instead of doing in the Presenter implementation of the respective SalesConfig and ProductConfig screens.